+++ tests/run.clj
@@ -73,6 +73,7 @@
 (require "tests/tasks_test")
 (require "tests/records_test")
 (require "tests/instant_template_test")
+(require "tests/data_readers_test")
 (require "tests/async_buffer_test")
 (require "tests/async_api_test")
 (require "tests/async_alts_test")